+++ src/screens/Addresses/ContactsScreen.tsx
@@ -19,7 +19,7 @@ along with the library. If not, see <http://www.gnu.org/licenses/>.
 import { StackScreenProps } from '@react-navigation/stack'
 import { ScrollViewProps } from 'react-native'
 
-import { ScrollScreenProps } from '~/components/layout/ScrollScreen'
+import ScrollScreen, { ScrollScreenProps } from '~/components/layout/ScrollScreen'
 import { useScrollEventHandler } from '~/contexts/ScrollContext'
 import { AddressTabsParamList } from '~/navigation/AddressesTabNavigation'
 import RootStackParamList from '~/navigation/rootStackRoutes'
@@ -35,12 +35,12 @@ const ContactsScreen = ({ navigation, style, ...props }: ScreenProps) => {
   const scrollHandler = useScrollEventHandler()
 
   return (
-    <ContactListScreenBase
-      onContactPress={(contactId: Contact['id']) => navigation.navigate('ContactScreen', { contactId })}
-      onNewContactPress={() => navigation.navigate('NewContactScreen')}
-      onScroll={scrollHandler}
-      {...props}
-    />
+    <ScrollScreen {...props} onScroll={scrollHandler}>
+      <ContactListScreenBase
+        onContactPress={(contactId: Contact['id']) => navigation.navigate('ContactScreen', { contactId })}
+        onNewContactPress={() => navigation.navigate('NewContactScreen')}
+      />
+    </ScrollScreen>
   )
 }
